+++ py_slides_term/candidates/__init__.py
@@ -7,8 +7,10 @@ from .filters import (
     BaseCandidateTermFilter,
     JapaneseConcatenationFilter,
     EnglishConcatenationFilter,
-    SymbolLikeFilter,
-    ProperNounFilter,
+    JapaneseSymbolLikeFilter,
+    EnglishSymbolLikeFilter,
+    JapaneseProperNounFilter,
+    EnglishProperNounFilter,
 )
 from .data import (
     PageCandidateTermList,
@@ -27,8 +29,10 @@ __all__ = [
     "BaseCandidateTermFilter",
     "JapaneseConcatenationFilter",
     "EnglishConcatenationFilter",
-    "SymbolLikeFilter",
-    "ProperNounFilter",
+    "JapaneseSymbolLikeFilter",
+    "EnglishSymbolLikeFilter",
+    "JapaneseProperNounFilter",
+    "EnglishProperNounFilter",
     "PageCandidateTermList",
     "PDFCandidateTermList",
     "DomainCandidateTermList",
