+++ py_slides_term/candidates/filters/term/__init__.py
@@ -1,12 +1,14 @@
 from .base import BaseCandidateTermFilter
 from .concatenation import JapaneseConcatenationFilter, EnglishConcatenationFilter
-from .symbollike import SymbolLikeFilter
-from .propernoun import ProperNounFilter
+from .symbollike import JapaneseSymbolLikeFilter, EnglishSymbolLikeFilter
+from .propernoun import JapaneseProperNounFilter, EnglishProperNounFilter
 
 __all__ = [
     "BaseCandidateTermFilter",
     "JapaneseConcatenationFilter",
     "EnglishConcatenationFilter",
-    "SymbolLikeFilter",
-    "ProperNounFilter",
+    "JapaneseSymbolLikeFilter",
+    "EnglishSymbolLikeFilter",
+    "JapaneseProperNounFilter",
+    "EnglishProperNounFilter",
 ]
