+++ py_slides_term/candidates/filters/term/concatenation/__init__.py
@@ -1,0 +1,7 @@
+from .japanese import JapaneseConcatenationFilter
+from .english import EnglishConcatenationFilter
+
+__all__ = [
+    "JapaneseConcatenationFilter",
+    "EnglishConcatenationFilter",
+]
