+++ py_slides_term/candidates/filters/term/propernoun/__init__.py
@@ -1,0 +1,7 @@
+from .japanese import JapaneseProperNounFilter
+from .english import EnglishProperNounFilter
+
+__all__ = [
+    "JapaneseProperNounFilter",
+    "EnglishProperNounFilter",
+]
