+++ py_slides_term/candidates/filters/term/propernoun/english.py
@@ -1,0 +1,11 @@
+from ..base import BaseEnglishCandidateTermFilter
+from py_slides_term.share.data import Term
+
+
+class EnglishProperNounFilter(BaseEnglishCandidateTermFilter):
+    # public
+    def __init__(self):
+        pass
+
+    def is_candidate(self, scoped_term: Term) -> bool:
+        return True
