+++ py_slides_term/endtoend/mappers/filters/term.py
@@ -5,8 +5,10 @@ from py_slides_term.candidates import (
     BaseCandidateTermFilter,
     JapaneseConcatenationFilter,
     EnglishConcatenationFilter,
-    SymbolLikeFilter,
-    ProperNounFilter,
+    JapaneseSymbolLikeFilter,
+    EnglishSymbolLikeFilter,
+    JapaneseProperNounFilter,
+    EnglishProperNounFilter,
 )
 
 
@@ -19,8 +21,10 @@ class CandidateTermFilterMapper(BaseMapper[Type[BaseCandidateTermFilter]]):
         term_filter_clses = [
             JapaneseConcatenationFilter,
             EnglishConcatenationFilter,
-            SymbolLikeFilter,
-            ProperNounFilter,
+            JapaneseSymbolLikeFilter,
+            EnglishSymbolLikeFilter,
+            JapaneseProperNounFilter,
+            EnglishProperNounFilter,
         ]
         for filter_cls in term_filter_clses:
             default_mapper.add(f"{module_path}.{filter_cls.__name__}", filter_cls)
