+++ py_slides_term/morphemes/__init__.py
@@ -1,5 +1,5 @@
 from .tokenizer import SpaCyTokenizer
-from .classifier import JapaneseMorphemeClassifier, EnglishMorphemeClassifier
+from .classifiers import JapaneseMorphemeClassifier, EnglishMorphemeClassifier
 from .data import BaseMorpheme, MorphemeSpaCyDic
 
 __all__ = [
