+++ py_slides_term/morphemes/classifiers/__init__.py
@@ -1,0 +1,7 @@
+from .japanese import JapaneseMorphemeClassifier
+from .english import EnglishMorphemeClassifier
+
+__all__ = [
+    "JapaneseMorphemeClassifier",
+    "EnglishMorphemeClassifier",
+]
