+++ py_slides_term/morphemes/classifiers/japanese.py
@@ -1,4 +1,4 @@
-from .data import BaseMorpheme
+from ..data import BaseMorpheme
 
 
 class JapaneseMorphemeClassifier:
@@ -11,12 +11,3 @@ class JapaneseMorphemeClassifier:
 
     def is_connector_symbol(self, morpheme: BaseMorpheme) -> bool:
         return morpheme.surface_form in {"・", "-"} and morpheme.pos == "補助記号"
-
-
-class EnglishMorphemeClassifier:
-    # public
-    def is_symbol(self, morpheme: BaseMorpheme) -> bool:
-        return morpheme.pos == "SYM"
-
-    def is_connector_symbol(self, morpheme: BaseMorpheme) -> bool:
-        return morpheme.surface_form == "-" and morpheme.pos == "SYM"
